add createFragmentSpecResolver
