+++ types/relay-runtime/lib/store/createFragmentSpecResolver.d.ts
@@ -1,0 +1,9 @@
+import { RelayContext, FragmentMap, Props, FragmentSpecResolver } from './RelayStoreTypes';
+
+export function createFragmentSpecResolver(
+    context: RelayContext,
+    containerName: string,
+    fragments: FragmentMap,
+    props: Props,
+    callback?: () => void,
+): FragmentSpecResolver;
